third_party/blink/web_tests/platform/linux/fast/forms/datetimelocal-multiple-fields/datetimelocal-multiple-fields-value-set-empty-expected.txt

Check if input.value="" clears an input with partially-specified value.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Empty text: mm/dd/yyyy, --:-- --
PASS eventSender.keyDown("ArrowUp"); getUserAgentShadowTextContent(input) is not emptyText
PASS input.value = ""; getUserAgentShadowTextContent(input) is emptyText
PASS successfullyParsed is true

TEST COMPLETE
